Fix a couple bad errors I introduced
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@21612 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Expressions/Management/avtExpressionEvaluatorFilter.C
@@ -674,7 +674,7 @@ avtExpressionEvaluatorFilter::ModifyContract(avtContract_p spec)
     }
 
     // Do these in the order the pipeline would.
-    for (size_t i = filters.size()-1 ; i >= 0 ; i--)
+    for (int i = (int)filters.size()-1 ; i >= 0 ; i--)
     {
         rv = filters[i]->ModifyContract(rv);
     }
